requirements.txt
# Core dependencies
pydantic>=2.0.0
pydantic-ai>=0.1.0
pydantic-ai-slim[groq]>=0.1.0
pyyaml>=6.0
httpx>=0.24.0

# Development dependencies
pytest>=7.0.0
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0